Trying to reduce memory usage.

+++ install/wfs_dump.sql
@@ -36,7 +36,7 @@ DROP TABLE IF EXISTS `sens`;
 CREATE TABLE `sens` (
   `id` mediumint(9) NOT NULL AUTO_INCREMENT,
   `temp` float DEFAULT NULL,
-  `hum` float DEFAULT NULL,
+  `hum` int DEFAULT NULL,
   `atp` int(11) DEFAULT NULL,
   `issame` int(2) DEFAULT NULL,
   `tmestmp` datetime DEFAULT CURRENT_TIMESTAMP,
